+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - support all standard EDM schema versions - Jakub Filak
 - Splits python representation of metadata and metadata parsing - Martin Miksik
 - Separate type repositories for individual versions of OData - Martin Miksik
+- Support for OData V4 primitive types - Martin Miksik
 
 ### Fixed
 - make sure configured error policies are applied for Annotations referencing